requirements/common.txt
alembic==1.13.1
bcrypt==4.1.2
blinker==1.7.0
cffi==1.16.0
click==8.1.7
colorama==0.4.6
cryptography==41.0.7
dnspython==2.4.2
dominate==2.9.1
email-validator==2.1.0.post1
Flask==3.0.0
Flask-Bcrypt==1.0.1
Flask-Bootstrap==3.3.7.1
Flask-CKEditor==0.5.1
Flask-Login==0.6.3
Flask-Mail==0.9.1
Flask-Migrate==4.0.5
Flask-Moment==1.0.5
Flask-PageDown==0.4.0
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.1
greenlet==3.0.3
idna==3.6
itsdangerous==2.1.2
Jinja2==3.1.3
jwt==1.3.1
Mako==1.3.0
MarkupSafe==2.1.3
packaging==23.2
psycopg2==2.9.9
pycountry==23.12.11
pycparser==2.21
python-dotenv==1.0.0
SQLAlchemy==2.0.25
typing_extensions==4.9.0
visitor==0.1.3
Werkzeug==3.0.1
WTForms==3.1.2